refactor(plugin): 解释器探测迁出 registry——独立 interpreter_env 模块
registry.rs 已超 3200 行，执行环境探测自成体系，整体迁至
src/interpreter_env.rs（含测试）：

- pub：ensure_interpreter_env（进程入口注入）
- pub(crate)：InterpreterEnv 快照、search_interpreter_program、
  user_home_dir（registry 重启 server 依赖 sidecar 处复用）
- lib.rs 顶层导出 ensure_interpreter_env，tiangong-app main() 调用
  路径随之上移；registry 内解析解释器改为引用新模块
- 顺带清理测试模块前重复的 #[cfg(test)] 属性
